Extrapolate TWC Config WW2 classes
Restructuring for flare work, and add general macro stuff whilst I'm
here
