+++ @twc_config_ww2/addons/twc_ww2_weapons/CfgAmmo.hpp
@@ -1,0 +1,9 @@
+class CfgAmmo{
+	class fow_e_no82;
+	class twc_no82: fow_e_no82{
+		hit = 100;
+		indirectHit = 325;
+		indirectHItRange = 2;
+		caliber = 4;
+	};
+};
