+++ @twc_config_ww2/addons/twc_ww2_weapons/CfgMagazines.hpp
@@ -1,0 +1,7 @@
+class CfgMagazines {
+	class fow_e_no82;
+	class twc_no82: fow_e_no82 {
+		ammo = "twc_no82";
+		displayName = "NO. 82 AT Grenade";
+	};
+};
